Check for takedowns only for recent destinations.
This prevents an issue where some things may not get flagged by all the providers and then the host takes it down. The host may take it down in a way that DNS resolution is still up but the scam has been taken down so nothing else will flag it. This reduces the severity of the issue by having the old scams eventually "fall off" and not get checked anymore.